Split r1cs/mod.rs
