+++ src/r1cs/linear_combination.rs
@@ -1,0 +1,55 @@
+//! Definition of linear combinations.
+
+use curve25519_dalek::scalar::Scalar;
+use std::iter::FromIterator;
+
+/// Represents a variable in a constraint system.
+#[derive(Copy, Clone, Debug)]
+pub enum Variable {
+    /// Represents an external input specified by a commitment.
+    Committed(usize),
+    /// Represents the left input of a multiplication gate.
+    MultiplierLeft(usize),
+    /// Represents the right input of a multiplication gate.
+    MultiplierRight(usize),
+    /// Represents the output of a multiplication gate.
+    MultiplierOutput(usize),
+    /// Represents the constant 1.
+    One(),
+}
+
+/// Represents a linear combination of
+/// [`Variables`](::r1cs::Variable).  Each term is represented by a
+/// `(Variable, Scalar)` pair.
+#[derive(Clone, Debug)]
+pub struct LinearCombination {
+    pub(super) terms: Vec<(Variable, Scalar)>,
+}
+
+impl Default for LinearCombination {
+    fn default() -> Self {
+        LinearCombination { terms: Vec::new() }
+    }
+}
+
+impl FromIterator<(Variable, Scalar)> for LinearCombination {
+    fn from_iter<T>(iter: T) -> Self
+    where
+        T: IntoIterator<Item = (Variable, Scalar)>,
+    {
+        LinearCombination {
+            terms: iter.into_iter().collect(),
+        }
+    }
+}
+
+impl<'a> FromIterator<&'a (Variable, Scalar)> for LinearCombination {
+    fn from_iter<T>(iter: T) -> Self
+    where
+        T: IntoIterator<Item = &'a (Variable, Scalar)>,
+    {
+        LinearCombination {
+            terms: iter.into_iter().cloned().collect(),
+        }
+    }
+}
